3154 Prevent overwriting munin-symlink cron file and print message instead

+++ installer/install_munin-symlink.sh
@@ -109,9 +109,15 @@ deploy_configuration() {
 setup_cron_job() {
     echo "Setting up cron job..."
     CRON_FILE="/etc/cron.d/munin-symlink"
-    sudo tee "$CRON_FILE" > /dev/null <<EOF
+
+    if ! sudo test -f "$CRON_FILE"; then
+        sudo tee "$CRON_FILE" > /dev/null <<EOF
 */5 * * * * root test -x /root/bin/munin-symlink.sh && /root/bin/munin-symlink.sh
 EOF
+    else
+        echo "Cron job already exists: $CRON_FILE"
+        echo "Skipping creation to preserve existing configuration."
+    fi
     sudo chmod 644 "$CRON_FILE"
     sudo chown root:root "$CRON_FILE"
 }
